ビデオ会議機能
タブが閉じられたときの処理作成中

+++ src/main/java/com/example/online_seminar/controller/SkywayController.java
@@ -1,6 +1,5 @@
 package com.example.online_seminar.controller;
 
-import com.example.online_seminar.entity.group.GroupMessage;
 import com.example.online_seminar.entity.group.Meeting;
 import com.example.online_seminar.entity.group.MeetingMember;
 import com.example.online_seminar.entity.user.User;
@@ -10,11 +9,10 @@ import com.example.online_seminar.repository.MeetingRepository;
 import com.example.online_seminar.repository.UserRepository;
 import org.springframework.security.core.Authentication;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
-
-import java.sql.Date;
-import java.text.SimpleDateFormat;
-import java.util.Calendar;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("meetings")
@@ -40,6 +38,8 @@ public class SkywayController {
     @PostMapping("/open")
     public void openMeeting(@RequestBody() String data, Model model, Authentication loginUser) {
 
+        System.out.println(data);
+
         String peerId = data.substring(data.indexOf("=") + 1, data.indexOf("&"));
         int groupId = Integer.parseInt(data.substring(data.lastIndexOf("=") + 1));
 
@@ -48,9 +48,9 @@ public class SkywayController {
         Meeting meeting = new Meeting();
         meeting.setGroupId(groupId);
         meeting.setUserName(loginUserName.getUserName());
+        meeting.setPeerId(peerId);
 
         meetingRepository.save(meeting);
-
 
         MeetingMember meetingMember = new MeetingMember();
         meetingMember.setGroupId(groupId);
@@ -58,31 +58,42 @@ public class SkywayController {
 
         meetingMemberRepository.save(meetingMember);
 
-        GroupMessage groupMessage = new GroupMessage();
+    }
 
-        Calendar calendar = Calendar.getInstance();
-        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
-        String sdfCalender = sdf.format((calendar.getTime()));
-        int groupMessageId = 0;
+    @PostMapping("/join")
+    public void joinMeeting(@RequestBody() String data, Model model, Authentication loginUser){
 
-        groupMessage.setGroupMessageId(groupMessageId);
-        groupMessage.setUserId(loginUser.getName());
-        groupMessage.setUserName(loginUserName.getUserName());
-        groupMessage.setCreateDatetime(Date.valueOf(sdfCalender));
-        groupMessage.setMessageContents("会議を開始しました。会議ID：" + peerId);
-        groupMessage.setGroupId(groupId);
+        System.out.println(data);
 
-        groupMessageRepository.save(groupMessage);
+        int groupId = Integer.parseInt(data.substring(data.lastIndexOf("=") + 1));
+
+        User loginUserName = userRepository.findByUserId(loginUser.getName());
+
+        MeetingMember meetingMember = new MeetingMember();
+        meetingMember.setGroupId(groupId);
+        meetingMember.setUserName(loginUserName.getUserName());
+
+        meetingMemberRepository.save(meetingMember);
 
     }
 
     @PostMapping("/close")
-    public void closeMeeting(@RequestBody String data) {
+    public void closeMeeting(@RequestBody String data, Authentication loginUser) {
+
+        System.out.println("削除前！！");
 
         int groupId = Integer.parseInt(data.substring(data.lastIndexOf("=") + 1));
+        User loginUserName = userRepository.findByUserId(loginUser.getName());
 
-        meetingMemberRepository.deleteAllByGroupId(groupId);
-        meetingRepository.deleteByGroupId(groupId);
+        System.out.println(loginUserName.getUserName());
+        meetingMemberRepository.deleteByUserName(loginUserName.getUserName());
+        if(meetingMemberRepository.countByGroupId(groupId) == 0) {
+            meetingRepository.deleteByGroupId(groupId);
+        }
+
+        System.out.println("削除した!!");
+//        meetingMemberRepository.deleteAllByGroupId(groupId);
+
 
 //        return "";
     }
